requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
asyncio>=3.4.3
SQLAlchemy==2.0.23 